more lib.rs docs
