+++ src/runner/runner.rs
@@ -5,6 +5,56 @@ use tokio::sync::Mutex;
 
 use tokio::prelude::*;
 
+/**
+The runner is the main "event loop" of this crate.
+
+It is created with a [Dispatcher][dispatcher]. It returns the new runner and the
+[Control][control] type.
+
+Once you're ready to start reading from the **Reader** and processing **Writes** you should call [Runner::run](#method.run).
+
+# Returns
+- A [`future`][future] which resolves to a [Status][status] once the Runner has finished.
+
+Interacting with the `Runner` is done via the [Control][control] type.
+
+# Example
+```rust
+# use tokio::spawn;
+# tokio::runtime::Runtime::new().unwrap().block_on(async {
+# let conn = tokio_test::io::Builder::new().wait(std::time::Duration::from_millis(10000)).build();
+use twitchchat::{Dispatcher, Status, Runner, RateLimit};
+// make a dispatcher
+let dispatcher = Dispatcher::new();
+// do stuff with the dispatcher (its clonable)
+// ..
+
+// create a new runner
+let (runner, control) = Runner::new(dispatcher, RateLimit::default());
+
+// spawn a task that kills the runner after some time
+let ctl = control.clone();
+spawn(async move {
+    // pretend some time has passed
+    ctl.stop()
+});
+
+// run, blocking the task.
+// you can spawn this in a task and await on that join handle if you prefer
+match runner.run(conn).await {
+    // for the doc test
+    Ok(Status::Canceled) => { assert!(true) }
+    Ok(Status::Eof) => { panic!("eof") }
+    Err(err) => { panic!("err") }
+};
+# });
+```
+
+[dispatcher]: ./struct.Dispatcher.html
+[control]: ./struct.Control.html
+[status]: ./enum.Status.html
+[future]: https://doc.rust-lang.org/std/future/trait.Future.html
+*/
 pub struct Runner {
     dispatcher: Dispatcher,
     receiver: Rx,
